extract datetime_funcs expressions to folders based on spark grouping
